+++ CHANGELOG.md
@@ -26,6 +26,7 @@ the list is now fixed inside the packaged logprep
 * refactored some processors to make use of the new helper methods
 * add `pre-commit` hooks to the repository, install new dev dependency and run `pre-commit install` in the root dir
 * the default `securityContext`for the pod is now configurable
+* allow `TimeParser` to get the current time with a specified timezone instead of always using local time and setting the timezone to UTC
 * remove `tldextract` dependency
 * remove `urlextract` dependency
 
@@ -34,6 +35,7 @@ the list is now fixed inside the packaged logprep
 * fix `confluent_kafka.store_offsets` if `last_valid_record` is `None`, can happen if a rebalancing happens
   before the first message was pulled.
 * fix pseudonymizer cache metrics not updated
+* fix incorrect timezones for log arrival time and delta time in input preprocessing
 * fix `_get_value` in `FilterExpression` so that keys don't match on values
 * fix `auto_rule_tester` to work with `LOGPREP_BYPASS_RULE_TREE` enabled